refactor: gather the app-facing driving triggers into conversation/driving.rs
