+++ src/conversation/handle.rs
@@ -443,79 +443,6 @@ where
         }
     }
 
-    /// Manually produce a Layer-3 recovery commit: mint our candidate and queue
-    /// its broadcast (only valid in recovery, where the steward gate is relaxed).
-    /// `Ok(true)` if a candidate was produced, `Ok(false)` if there was nothing
-    /// to commit or we already hold a local candidate this round — so the app
-    /// can call it every cycle in recovery without minting duplicates. de-mls no
-    /// longer auto-mints; the app owns the recovery-commit timing.
-    pub fn commit_in_recovery<Pr>(
-        &mut self,
-        provider: &Pr,
-        signer: &impl Signer,
-    ) -> Result<bool, ConversationError>
-    where
-        Pr: OpenMlsProvider,
-        <Pr::StorageProvider as StorageProvider<1>>::Error: StdError + Send + Sync + 'static,
-    {
-        if !self.is_in_recovery_mode() {
-            return Err(ConversationError::NotInRecovery);
-        }
-        if self.queues.commit_round.has_local_candidate() {
-            return Ok(false);
-        }
-        self.mint_and_broadcast_candidate(provider, signer)
-    }
-
-    /// Start the commit round now instead of waiting out a commit-inactivity
-    /// timer: enter `Freezing` and, if we're a steward, mint our candidate from
-    /// the approved batch. de-mls no longer keeps this timer — the app calls
-    /// this when its own liveness policy decides the epoch steward has sat on
-    /// approved work long enough, or when a backup takes over a silent primary.
-    ///
-    /// `Ok(true)` when the freeze was entered, `Ok(false)` when there is nothing
-    /// to do: not in `Working`, no approved work, or an election in flight
-    /// (committing on a stale steward list only produces a NoCandidate). Mirrors
-    /// the guards of [`Self::pending_commit_work`].
-    pub fn commit_now<Pr>(
-        &mut self,
-        provider: &Pr,
-        signer: &impl Signer,
-    ) -> Result<bool, ConversationError>
-    where
-        Pr: OpenMlsProvider,
-        <Pr::StorageProvider as StorageProvider<1>>::Error: StdError + Send + Sync + 'static,
-    {
-        if self.current_state() != ConversationState::Working
-            || self.queues.approved_proposals_count() == 0
-            || self.queues.has_election_in_flight()
-        {
-            return Ok(false);
-        }
-        let Some(event) = self.start_freezing() else {
-            return Ok(false);
-        };
-        self.on_freeze_entered(provider, signer, event)?;
-        Ok(true)
-    }
-
-    /// Re-broadcast our commit candidate for the current round, so members that
-    /// dropped the original receive it and apply it — the steward's alternative
-    /// to survivors minting a competing commit, which forks. `Ok(true)` if a
-    /// candidate was re-sent, `Ok(false)` when we hold none for this round
-    /// (nothing minted yet, or it already merged and cleared the buffer). de-mls
-    /// does not time this; the app calls it when its liveness policy judges the
-    /// commit was likely dropped.
-    pub fn resend_commit(&self) -> Result<bool, ConversationError> {
-        match self.queues.commit_round.local_candidate_bytes() {
-            Some(payload) => {
-                self.broadcast(payload);
-                Ok(true)
-            }
-            None => Ok(false),
-        }
-    }
-
     /// Finalize the active commit round.
     pub(crate) fn finalize_commit_round<Pr>(
         &mut self,
